Enable building on Windows
An attempt to install the package on Windows (after removing the OS
restriction) resulted in the following warning:

datetime.c:102:14: warning: 'tzname' redeclared without dllimport attribute: previous dllimport ignored [-Wattributes]
 extern char *tzname[2];
               ^

The fix is to declare tzname with the dllimport attribute, as was done
for Cygwin.

While researching how R itself configures this block of code, I learned
that USE_INTERNAL_MKTIME is set by default. It is unset if configure
determines that the system tools behave reasonably. It's set by default
on MacOS, and it's what is used on Windows.